requirements.txt
nltk
scikit-learn 
networkx 
gensim 
numpy 
PyPDF2 
python-pptx 
streamlit